+++ packages/bridge-ts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.277-alpha.0](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/bridge-ts@1.0.276...@injectivelabs/bridge-ts@1.0.277-alpha.0) (2023-02-20)
+
+**Note:** Version bump only for package @injectivelabs/bridge-ts
+
+
+
+
+
 ## [1.0.276](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/bridge-ts@1.0.275...@injectivelabs/bridge-ts@1.0.276) (2023-02-18)
 
 **Note:** Version bump only for package @injectivelabs/bridge-ts